[VPN 2.0] Split VPN utils into shared and V1-only targets
To add a new BraveVpnService implementation based on Architecture 2.0,
which must co-exist with Architecture 1.0 for quite a while, we need
to split service's interface and implementation. All the external
components will keep accessing VPN service via the BraveVpnService
interface, but the implementation mostly goes into BraveVpnServiceImpl.

This incremental change splits "brave_vpn" target into two:
- "brave_vpn" itself, shared between V1 and V2 implementations.
- "v1_utils", a strictly V1 target, currently containing the VPN
  connection manager utility code for Architecture 1.0.

Most existing targets will continue depending on the shared
"brave_vpn"; "v1_utils" is used only in the specific place where V1
connection manager creation is done, and will be later gated by a
buildflag.

Implements part of https://github.com/brave/brave-browser/issues/54597
