+++ browser/brave_vpn/sources.gni
@@ -42,6 +42,7 @@ if (enable_brave_vpn) {
   brave_browser_brave_vpn_deps += [
     "//base",
     "//brave/browser/brave_vpn",
+    "//brave/browser/brave_vpn:v1_utils",
     "//brave/components/brave_vpn/browser",
     "//chrome/browser/profiles:profile",
     "//components/keyed_service/content",
