+++ browser/brave_vpn/vpn_utils.h
@@ -6,29 +6,14 @@
 #ifndef BRAVE_BROWSER_BRAVE_VPN_VPN_UTILS_H_
 #define BRAVE_BROWSER_BRAVE_VPN_VPN_UTILS_H_
 
-#include <memory>
-
-#include "base/memory/scoped_refptr.h"
-
 namespace content {
 class BrowserContext;
 }  // namespace content
 
-namespace network {
-class SharedURLLoaderFactory;
-}  // namespace network
-
-class PrefService;
-
 namespace brave_vpn {
-
-class BraveVPNConnectionManager;
 
 bool IsBraveVPNEnabled(content::BrowserContext* context);
 bool IsAllowedForContext(content::BrowserContext* context);
-std::unique_ptr<BraveVPNConnectionManager> CreateBraveVPNConnectionManager(
-    scoped_refptr<network::SharedURLLoaderFactory> url_loader_factory,
-    PrefService* local_prefs);
 
 }  // namespace brave_vpn
